requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
httpx
PyJWT
slowapi
pydantic-settings
pytest
pytest-asyncio
